requirements.txt
Flask~=3.0.3
firebase-admin
Flask-Mail

python-dotenv~=1.0.1
Pyrebase~=3.0.27
Werkzeug~=3.0.3
itsdangerous~=2.2.0